Move short conv to attn_gym.linear.short_conv
## Human Note

## Agent note

The fused causal short convolution is model-agnostic (KDA and GDN serving both use it), so move
it from attn_gym/linear/kda/short_conv to attn_gym/linear/short_conv and extract the torch-only
operator schemas into short_conv/ops.py. The public names were already exported from
attn_gym.linear; attn_gym.linear.kda keeps lazy re-exports for backward compatibility.

## Test Plan

```bash
gpu-run auto -- .venv/bin/pytest -q -n 6 test/test_short_conv_cute.py
```

stack-info: PR: https://github.com/meta-pytorch/attention-gym/pull/390, branch: drisspg/stack/98
