+++ attn_gym/linear/short_conv/activations.py
